+++ src/main/java/com/braintreegateway/PaymentMethodGateway.java
@@ -37,6 +37,18 @@ public class PaymentMethodGateway {
         return parseResponse(response).getTarget();
     }
 
+    public Result<PaymentMethodNonce> grant(String token) {
+        String request = new RequestBuilder("payment-method").addElement("shared-payment-method-token", token).toXML();
+        NodeWrapper response = http.post(configuration.getMerchantPath() + "/payment_methods/grant", request);
+        return new Result<PaymentMethodNonce>(response, PaymentMethodNonce.class);
+    }
+
+    public NodeWrapper revoke(String token) {
+        String request = new RequestBuilder("payment-method").addElement("shared-payment-method-token", token).toXML();
+        NodeWrapper response = http.post(configuration.getMerchantPath() + "/payment_methods/revoke", request);
+        return response;
+    }
+
     public Result<? extends PaymentMethod> parseResponse(NodeWrapper response) {
         if (response.getElementName() == "paypal-account") {
             return new Result<PayPalAccount>(response, PayPalAccount.class);